+++ src/controllers/VerifyController.php
@@ -5,7 +5,7 @@ namespace born05\twofactorauthentication\controllers;
 use Craft;
 use craft\web\Controller;
 use craft\web\View;
-use craft\elements\User;
+use craft\web\User;
 use born05\twofactorauthentication\Plugin as TwoFactorAuth;
 use born05\twofactorauthentication\web\assets\verify\VerifyAsset;
 
@@ -37,25 +37,30 @@ class VerifyController extends Controller
         $authenticationCode = $request->getBodyParam('authenticationCode');
 
         // Get the current user
-        $user = Craft::$app->getUser()->getIdentity();
+        /** @var User */
+        $userSession = Craft::$app->getUser();
+        $user = $userSession->getIdentity();
 
         if (TwoFactorAuth::$plugin->verify->verify($user, $authenticationCode)) {
             // Get the session duration
-            $sessionDuration = Craft::$app->getUser()->getRemainingSessionTime();
+            $sessionDuration = $userSession->getRemainingSessionTime();
 
             // Throw the after login event, because we blocked it earlier for non-cookieBased events.
             $this->afterLogin($user, false, $sessionDuration);
 
             return $this->_handleSuccessfulLogin($user);
         } else {
+            $errorCode = \craft\elements\User::AUTH_INVALID_CREDENTIALS;
+            $errorMessage = Craft::t('two-factor-authentication', 'Authentication code is invalid.');
+
             return $this->asFailure(
-                Craft::t('two-factor-authentication', 'Authentication code is invalid.'),
+                $errorMessage,
                 data: [
-                    'errorCode' => User::AUTH_INVALID_CREDENTIALS,
+                    'errorCode' => $errorCode,
                 ],
                 routeParams: [
-                    'errorCode' => User::AUTH_INVALID_CREDENTIALS,
-                    'errorMessage' => Craft::t('two-factor-authentication', 'Authentication code is invalid.'),
+                    'errorCode' => $errorCode,
+                    'errorMessage' => $errorMessage,
                 ]
             );
         }
@@ -69,9 +74,10 @@ class VerifyController extends Controller
      *
      * @return Response
      */
-    private function _handleSuccessfulLogin(User $user): Response
+    private function _handleSuccessfulLogin(\craft\elements\User $user): Response
     {
         // Get the return URL
+        /** @var User */
         $userSession = Craft::$app->getUser();
         $returnUrl = $userSession->getReturnUrl();